Added integral functionality to cdf lib
