+++ packages/cdf/continuousDistribution.js
@@ -1,5 +1,5 @@
-
 const { interpolate, range, min, max } = require('./functions');
+const _ = require('lodash');
 
 class ContinuousDistribution {
   /**
@@ -146,6 +146,25 @@ class ContinuousDistribution {
    */
   sample(size) {
     return Array.from(Array(size), () => this.sampleSingle());
+  }
+
+  /**
+   * Finds the integral. Takes the average Y value between points, treating them like a triangle.
+   * @return {number[]}
+   */
+  integral(){
+    let integral = 0
+    for (let i = 1; i < this.ys.length; i++) {
+      let thisY = this.ys[i];
+      let lastY = this.ys[i-1];
+      let thisX = this.xs[i];
+      let lastX = this.xs[i-1];
+      if (_.isFinite(thisY) && _.isFinite(lastY) && _.isFinite(thisX) && _.isFinite(lastX)){
+        let sectionInterval = ((thisY+lastY)/2)*(thisX-lastX);
+        integral = integral + sectionInterval
+      }
+    }
+    return integral
   }
 }
