+++ packages/cdf/continuousDistributionCombination.js
@@ -45,9 +45,7 @@ class ContinuousDistributionCombination {
    */
   combineYsWithFn(sampleCount, fn) {
     const xs = this.equallyDividedXs(sampleCount);
-    console.log("Initial Points", xs.map(xPoint => (this.allYsAtXPoint(xPoint))));
     const result = xs.map(xPoint => fn(this.allYsAtXPoint(xPoint)));
-    console.log("Calculations", result)
     return new ContinuousDistribution(xs, result);
   }
